requirements.txt
crewai[tools,agentops]==0.114.0
scrapegraph-py
tavily-python
pydantic
